move signal handlers to receivers.py
